\name{seasonal}
\docType{data}
\alias{seasonal}
\title{Simulated data with periodic component}
\description{This data give the 1000 simulated observations with survival time and a periodic (or seasonal) component time scale. 
One addditional covariate is simulated too.
}
\usage{data(seasonal)
}
\format{A data frame containing 1000 sampled observations. Description of columns:
\itemize{
\item{birth}{Simulated seasonal time of entry in the study or starting point, like season of birth in a year.}
\item{surv}{Simulated survival time.}
\item{status}{Simulated censoring indicator in the usual notation for right censored data.}
\item{covariate}{Simulated covariate.}
}
}
\keyword{datasets}
