\name{TwoWaySurvival-internal}
\alias{TwoWayVarCoefBackfit}
\title{Internal function for fitting of the two-way survival model}
\description{The function 'TwoWayVarCoefBackfit' implements an optimization procedure and 
is used internally in the body of the 'TwoWaySurvfitCreate' function.}
\details{This function is not to be called by the user.}
\keyword{internal}

