\name{TwoWaySurvival-internal}
\alias{BackfitTruncated}
\alias{BackfitBspline}
\title{Internal functions for fitting of the two-way survival model}
\description{The function 'BackfitTruncated' implements an optimization procedure in the case of truncated polynomials
and is used internally in the body of the 'TwoWaySurvfitCreate' function. The same is for the function 'BackfitBspline' in the case of
B-splines as a model basis.}
\details{This functions are not to be called by the user.}
\keyword{internal}

