\name{TwoWaySurvival-internal}
\alias{BackfitTruncated}
\alias{BackfitBspline}
\alias{BsplineOneway}
\alias{TruncatedOneway}
\title{Internal functions for fitting of the one- and two-way survival model}
\description{The function 'BackfitTruncated' implements an optimization procedure in the case of truncated polynomials
and is used internally in the body of the 'TwoWaySurvfitCreate' function. The same is for the function 'BackfitBspline' in the case of
B-splines as a model basis. Complementary the two other functions, 'TruncatedOneway' and 'BsplineOneway', 
are attached for fitting a one dimensional model with accordingly chosen approximation bases.}
\details{This functions are not to be called by the user.}
\keyword{internal}

